refactor(dashboard): extract keeper-detail-panels and keeper-detail-runtime from keeper-detail.ts
Split keeper-detail.ts (794 LOC) into three modules under 400 lines each:
- keeper-detail.ts (251): main overlay, global state, comms panel
- keeper-detail-panels.ts (340): KPIs, charts, field dictionary, TRPG, equipment, relationships, traits, autonomy
- keeper-detail-runtime.ts (245): runtime signals, neighborhood, tool audit
